refactor: 轮次写面收形——Settle 交回话题、append 按场景验键并返回槽位；字段面审计净删四处
Why: 宿主自报的轮次键此前不验，拼错或自造的 topic id 会把内容写进任何读取都列不出的孤儿键；Update 蒸出的关键词轨要宿主再读一次才拿得到；公开 JSON 的 id_hash 命名与门面「id 即 hex」不一致；逐层字段审计另发现 MBTI 类型词与四轴双存可漂移、画像 payload 的 IDHash 零读者、L1 数值链 float32/float64 混用、读侧时间戳单位只在 L4 写明

- Update 改名 Settle 并交回落盘后的话题（关键词轨随调用返回）；AppendArchive 改为 (sceneID, topicID) 成对验键、返回占用的 Seq；MCP append 收成一批一次往返，中途失败已落的 seqs 随错误返回
- PlanCreate 折掉（parentSeq 0 即开树），公开面钉为 25 + 7；DB.Stats 诊断口（文件字节 + 可达记录数），MCP 新增 memhop_compact（临时名锁在 db-dir 内，先关库静默再换名重开）与 memhop_status 两个读数，工具面钉 25 个
- 公开 DTO 的 id_hash 统一改名 id；SceneNodeView 暴露 EmotionSet；trajectory_read 入参 session_id 改 topic_id
- 保留窗可配：MemHopDefaults.ContentRetentionMs（默认 7 天，无「永不过期」写法）
- MBTI.Type 不再落盘（json:"-"），ReadProfileSlot 从四轴重派生（core.DeriveMBTIType），两处自带漂移的夹具改成自洽；ProfileSlot.IDHash 删除
- L1 数值链统一 float64（Importance/Weight/阈值/jaccard/蒸馏样本/api SceneNodeView）
- 公开面其余时间戳字段写明毫秒单位；README 双语跟上本轮写面（此前仍写 Update/PlanCreate/26+24）；content/config/dream/turn 四包 agent.md 同步，cap 与 repo 两包核对无失同步

+++ internal/cap/engram/decay.go
@@ -20,9 +20,9 @@ import (
 type DecayParams struct {
 	LambdaNode             float64
 	LambdaEdge             float64
-	NodeRemoveThreshold    float32
-	NodePruneEdgeThreshold float32
-	EdgeRemoveThreshold    float32
+	NodeRemoveThreshold    float64
+	NodePruneEdgeThreshold float64
+	EdgeRemoveThreshold    float64
 	MinEdgeNodes           int
 }
 
@@ -170,7 +170,7 @@ func skipDeepNode(node *core.SceneNode, l2Meta *index.L2MetaIndex) bool {
 func decayOneNode(engine *core.StorageEngine, agentID uint64, cfg *DecayParams, node *core.SceneNode, nowMs int64, report *DecayReport, removedNodeIDs map[uint64]bool, clearedEdges map[uint64]map[uint64]bool) error {
 	dtHours := common.ElapsedHours(nowMs, node.UpdatedAt)
 	lambda := applyEmotionalBoost(cfg.LambdaNode, node.Valence, node.Arousal)
-	newImportance := node.Importance * float32(math.Exp(-lambda*dtHours))
+	newImportance := node.Importance * math.Exp(-lambda*dtHours)
 	if newImportance < cfg.NodeRemoveThreshold {
 		if _, err := engine.DeleteRecord(agentID, node.IDHash); err != nil {
 			return err
@@ -244,7 +244,7 @@ func decayOneEdge(engine *core.StorageEngine, agentID uint64, cfg *DecayParams, 
 		baseMs = edge.CreatedAt
 	}
 	dtHours := common.ElapsedHours(nowMs, baseMs)
-	newWeight := edge.Weight * float32(math.Exp(-cfg.LambdaEdge*dtHours))
+	newWeight := edge.Weight * math.Exp(-cfg.LambdaEdge*dtHours)
 
 	// A member is gone either because this pass removed it or because something
 	// outside it did — a scene delete takes its node with it between two Dreams.
